[Izza] Operation - Change parameter for Operation to Long

+++ app/src/main/java/id/co/iconpln/controlflowapp/Operation.kt
@@ -1,8 +1,8 @@
 package id.co.iconpln.controlflowapp
 
 sealed class Operation {
-    class Add(val value: Int) : Operation()
-    class Divide(val value: Int) : Operation()
-    class Multiply(val value: Int) : Operation()
-    class Substract(val value: Int) : Operation()
+    class Add(val value: Long) : Operation()
+    class Divide(val value: Long) : Operation()
+    class Multiply(val value: Long) : Operation()
+    class Substract(val value: Long) : Operation()
 }
